requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-GraphQL==2.0.0
Flask-JWT-Extended==3.21.0
Flask-SQLAlchemy==2.4.0
Flask-Migrate==2.5.2
graphene==2.1.7
graphene-sqlalchemy==2.2.1
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
PyJWT==1.7.1
pytz
python-decouple==3.1
psycopg2-binary==2.8.3
SQLAlchemy==1.3.6
uWSGI==2.0.18
Werkzeug==0.15.5
